==> 5_hold.ipynb <==
import tushare as ts
import pandas as pd
import numpy as np
import os
import time
import tqdm
import talib as ta
import mplfinance as mpf
import matplotlib.pyplot as plt
from datetime import datetime
from pathlib import Path
import shutil

# TOEKN_PATH = os.path.expanduser("~/.tushare.token")

# with open(TOEKN_PATH, "r") as f:
#     token = f.read().strip()
#     ts.set_token(token=token)
#     pro = ts.pro_api(token=token)


stock_basic_df = pd.read_csv("./data/stock_basic_df.csv")
ggt_df = pd.read_csv("./data/ggt_basic_df.csv")
etf_df = pd.read_csv("./data/etf_basic_df.csv")
basic_df = pd.read_csv("./data/basic_df.csv")
industry_df = pd.read_csv("./data/industry.csv")
buy_list = pd.read_csv("./data/buy_list.csv")['Column1'].tolist()
up_list = pd.read_csv("./data/up_list.csv")['Column1'].tolist()

# get_industry_info
hold_csv = 'table1120.csv'
hold_df = pd.read_csv(hold_csv)
hold_df = hold_df['证券代码'].apply(lambda x: str(x).zfill(6))
hold_df = hold_df.apply(lambda x: x if not isinstance(x, str) else (x + '.SH' if (x.startswith('6') or x.startswith('5')) else (x + '.SZ' if (x.startswith('0') or x.startswith('1') or x.startswith('3')) else x)))
hold_list = list(hold_df)
hold_list[0]

sell_list = [i for i in hold_list if i not in up_list]
sell_list

hold_df = pd.DataFrame()

for ts_code in hold_list:
    # 从 daily 里读 csv,只读第一行然后按 circ_mv 排序
    #df = pd.read_csv(os.path.join('./data/daily', f"{ts_code}.csv"))
    
    if ts_code not in basic_df['ts_code'].values:
        hold_df.loc[ts_code, 'name'] = str(ts_code)
        #hold_df.loc[ts_code, 'industry'] = pd.NA
        continue
    name = basic_df.loc[basic_df['ts_code'] == ts_code, 'name'].values[0]
    industry = basic_df.loc[basic_df['ts_code'] == ts_code, 'industry'].values[0]
    hold_df.loc[ts_code, 'name'] = name
    hold_df.loc[ts_code, 'industry'] = industry
print(hold_df.sort_values(by='industry'))
# 计算industry列中每个值出现的次数
industry_counts = hold_df['industry'].value_counts()
# 找出出现次数大于1的行业，即重复项
duplicates = industry_counts[industry_counts > 1].index.tolist()
duplicates

t_code = '600089.SH'
name = basic_df.loc[basic_df['ts_code'] == t_code, 'name'].values[0]
industry = basic_df.loc[basic_df['ts_code'] == t_code, 'industry'].values[0]
print(t_code,name,industry)

t_df = hold_df[hold_df['industry'] == industry]
t_df

print(len(hold_df))
hold_df.sort_index()
